Update how_can_i_use_the_ivar_uvar_wvar_and_cvar_options_to_precisely_control_the_permutations.md
This is based on an experience where using cvar led to a homogeneous permutation distribution.  I had not understood what was meant by permuting between blocks and was surprised to be always getting maximum or minimum p-values.  On Arjen Stolk's advice I had a look at the resampledesign function and it seems like it permutes over non-unique blocks/observations, and because my data were all unique it was using only a single set of indices for all permutations.